Restructured and fixed small typo
