+++ articles/00FF0000/analysis/att1/bb97dc4c.csv
