+++ articles/00FF0000/validation.py
